+++ backend/src/app.module.ts
@@ -5,6 +5,7 @@ import { PollModule } from './poll/poll.module';
 import { VoteGateway } from './vote/vote.gateway';
 import { QuestionModule } from './question/question.module';
 import { AnswerModule } from './answer/answer.module';
+import { VoteModule } from './vote/vote.module';
 
 @Module({
   imports: [
@@ -12,8 +13,9 @@ import { AnswerModule } from './answer/answer.module';
     PollModule,
     QuestionModule,
     AnswerModule,
+    VoteModule,
   ],
   controllers: [],
-  providers: [VoteGateway],
+  providers: [],
 })
 export class AppModule {}
