+++ backend/src/vote/vote.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { VoteGateway } from './vote.gateway';
+
+@Module({
+  providers: [VoteGateway],
+  exports: [VoteGateway],
+})
+export class VoteModule {}
